requirements.txt
openai>=1.40.0
pypdf>=4.2.0
python-dotenv>=1.0.1
tqdm>=4.66.4
together>=1.2.0
